Create useImpersonatedServiceAccount hook
